// MusicSpotClient/src/app/shared/enums/enums.ts
export enum ArchitectureTypeEnum{
    Monolith = 0,
    Microservices = 1
}


export enum GenreEnum{
    Rock = 0,
    Pop = 1,
    Indie = 2,
    Metal = 3,
    Classical = 4,
    House = 5,
    HipHop = 6,
    Jazz = 7,
    Rap = 8,
    Raggae=9,
    Punk = 10,
    Disco = 11,
    Electronica = 12,
    Dance = 13,
    Techno = 14,
    Folk = 15
}

export enum DecadeEnum{
    D2010_2020 = 0,
    D2000_2010 = 1,
    D1990_2000 = 3,
    D1980_1990 = 4,
    D1970_1980 = 5,
    D1960_1970 = 6,
    D1950_1960 = 7
}